Add files via upload
Class that can process the lamp, cell and dark data in tandem or separately.  The pixel space data is turned into integer 
quantities to better reflect the reality of pixels on the detector.  The data is then filtered for only the regions on the 
detector that contain spectrum.  The y-axis is then turned into counts with zero counts filled in to x-pixel locations where 
there is no data.  This is all put into a csv that also contains the integration time.  Both 2D and 1D histogram plots
are generated for the data sets that are processed.