Add files via upload
RANSAC model applied to tilted spectral lines in order to straighten the tilt.  This file specifically applies the RANSAC model to sparse data sets.  The RANSAC model should have already be run on a dense data set and then stored as a file that would then be applied to these spares data sets.  The file that runs the RANSAC model on the dense data set is "TraceLymanalpha.py"