Add files via upload
Ransac Model calibration is based off of the 05_09_2024_lamp_test8 data.  The file "WavelengthFWHM_Bright" is only for that data set and is intended to find the fwhm of the lyman alpha line for the resolution calculation.  The file 'ransac_model.pkl' is the model based off of the test8 data.  If you don't have this file in your directory it will retrain the model.  After you apply the ransac model to the data set, then you can run the wavelength calibration code and then do whatever else you want.